requirements.txt
py-cord==2.5.0
sqlite==3.45.3
python-dotenv~=0.21.0
tabulate~=0.9.0